requirements.txt
akinator>=1.1.1
antlr4-python3-runtime>=4.13.1
asyncio>=3.4.3
async-timeout>=4.0.3
beautifulsoup4>=4.12.2
blend-modes>=2.1.0
bs4>=0.0.2
c2pa-python>=0.4.0
CherryPy>=18.9.0
colormath>=3.0.0
cython>=3.0.10
discord>=2.3.2
filetype>=1.2.0
free-proxy>=1.1.1
gmpy2>=2.2.0a2
googletrans==4.0.0rc1
httpx>=0.27.0
imagesize>=1.4.1
inspiro>=0.0.3
knackpy>=1.1.1
markdownify>=0.12.1
matplotlib>=3.8.1
mpmath>=1.3.0
nekos.py>=1.1.0
numpy>=1.26.3
orjson>=3.10.3
pdf2image>=1.17.0
pillow>=10.3.0
pillow-heif>=0.16.0
prettytable>=3.9.0
psutil>=5.9.8
py-cpuinfo>=9.0.0
Pybooru>=4.2.2
pygame>=2.5.2
PyNaCl>=1.5.0
pynvml>=11.5.0
pypng>=0.20220715.0
pyqrcode>=1.2.1
python-dateutil>=2.8.2
pytz>=2024.1
requests>=2.31.0
rule34>=1.8.1
selenium>=4.19.0
sympy>=1.12
tiktoken>=0.6.0
tsc_utils>=0.2.5
typing_extensions>=4.9.0
wand>=0.6.13
yt-dlp>=2024.4.9